for WH, have to permute 'z' to '1'
The problem seems to be that our way of looking at the boundary doesn't make
sense if the boundary doesn't have the same degree on all terms.  We ignore
the boundary in the 'z' direction because its terms don't have degree 4.
The solution is to make sure that 'z' is in the interior, not the boundary.